requirements.txt
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.2
Flask-Migrate==2.0.3
Flask-RESTful==0.3.6
Flask-Script==2.0.5
Flask-SQLAlchemy==2.2
gunicorn==19.7.1
flask-bootstrap=3.3.7.1